+++ src/components/suggest/index.ts
@@ -6,8 +6,6 @@
  * Side Public License, v 1.
  */
 
-export { EuiSuggestInput, EuiSuggestInputProps } from './suggest_input';
-
 export { EuiSuggestItem, EuiSuggestItemProps } from './suggest_item';
 
 export { EuiSuggest, EuiSuggestProps, EuiSuggestionProps } from './suggest';